Refactor semantic assembly module into smaller, cohesive units
- Extracted `Assembler` struct and core API to `src/semantic/assembly/core.rs`.
- Extracted validation logic (agreement, limits, markers) to `src/semantic/assembly/checks.rs`.
- Extracted POS handling logic (nominal, verb, etc.) to `src/semantic/assembly/handlers.rs`.
- Moved unit tests to `src/semantic/assembly/tests.rs`.
- Updated `src/semantic/assembly/mod.rs` to act as a facade re-exporting the new modules.
- Added entry to `.jules/atlas.md`.

This breaks the "Blob" anti-pattern in `src/semantic/assembly/mod.rs`, improving maintainability and testability.
